pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "proxy_project"
version = "0.1.0"
description = "Django REST API proxy service for Polygon.io with WebSocket support"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [
    { name = "iklo" }
]
dependencies = [
    "django>=4.2.0",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "channels>=4.0.0",
    "channels-redis>=4.1.0",
    "polygon-api-client>=1.12.3",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.27.0",
    "redis>=5.0.0",
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    "httpx>=0.28.1",
    "daphne>=4.2.0",
    "django-cors-headers>=4.7.0",
    "django-redis>=5.4.0",
    "requests>=2.32.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=24.0.0",
    "flake8>=7.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["proxy_project"] 
